rag_flask/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
elastic-transport==8.15.0
elasticsearch==8.15.0
filelock==3.16.0
Flask==3.0.3
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
openai==1.44.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1+cpu
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uuid==1.30
Werkzeug==3.0.4
